Decompose Foundation interfaces, extract L2 handlers, split ServiceDispatcher
Reorganize F_platform/interfaces/ into three semantic subdirectories:
- dispatch/ -- 12 domain-specific dispatcher interfaces (motion, encoder, etc.)
- hal/ -- hardware abstraction interfaces (transport, SPI, motor driver, etc.)
- types/ -- shared data types (telemetry, events, velocity quality, etc.)

Extract all command handlers from monolithic command_parser.cpp into
handlers/*.cpp (motion, system, control, encoder, driver, config, display,
queue, safety). command_parser.cpp now contains only dispatch tables and
response helpers.

Split ServiceDispatcher into per-domain sd_*.cpp files to match the 12
decomposed interfaces, reducing the monolithic wiring header.

Add transport HAL interfaces (IRttChannel, IUartBus) with L4/L5 driver
implementations, and new traced wrappers (traced_lock, traced_spi_bus).

+++ server/layers/L3_services/infra/trace.cpp
@@ -40,7 +40,7 @@ void record(Dir d, const char* tag, uint32_t arg0) {
     s_total++;
 }
 
-void recordIface(uint8_t boundary, const char* label, const char* method,
+void recordIface(uint16_t boundary, const char* label, const char* method,
                  uint32_t arg0, const char* detail) {
     InterruptGuard guard;
